feat(swagger): fix bookings and swagger

+++ backend/controllers/BookingController.js
@@ -3,12 +3,13 @@ const { firestore } = require('../firebase/firebase');
 const createBooking = async (req, res) => {
     const { 
         roomNum, 
-        userId, 
         fromTime, 
         toTime, 
         status,
         paymentMethod 
     } = req.body;
+
+    const { uid: userId } = req.user;
 
     try {
         if (!roomNum || !userId || !fromTime || !toTime || !status || !paymentMethod){
@@ -38,11 +39,17 @@ const deleteBooking = async (req, res) => {
         bookingId
     } = req.body
 
+    const { uid: userId } = req.user;
+
     try {
         const doc = await firestore.collection('bookings').doc(bookingId).get();
         const id = doc.id;
         if (!doc.exists) {
             return res.status(400).json({ error: 'Bad request' });
+        }
+
+        if (doc.data().userId !== userId) {
+            return res.status(403).json({ error: 'No permission' });
         }
 
         await doc.ref.delete();
